+++ NEWS.md
@@ -1,6 +1,10 @@
 # argonR 0.2.0.9000
 - add hex
-- add a view argument to `argonPageTemplate()`.
+- add a view argument to `argonPageTemplate()`
+- add src_collapsed arg to `argonNavbar()` so that brand logo are displayed on small screens
+- add `argonContact()`
+- add favicon argument to `argonPage()` and update assets folder accordingly
+- add rounded argument to `argonButton()`
 
 # argonR 0.1.0
 - first release